job-market-skill-gap-analyzer/requirements.txt
# Core
python-dotenv==1.0.0
PyYAML==6.0.1

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0

# NLP
spacy==3.7.2
pandas==2.1.4
python-dateutil==2.8.2

# Big Data
pyspark==3.5.0

# Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
supabase==2.3.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Frontend
streamlit==1.30.0
plotly==5.18.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
